refactoring: move log formatter to a separate module
