+++ src/config/config.py
@@ -18,16 +18,6 @@ PYTHON_MIN = (3, 14)
 PYTHON_MAX = (4, 0)
 
 
-class NoTraceExceptionFormatter(logging.StreamHandler):
-    """Logging formatter without traceback for exceptions"""
-    def format(self, record):
-        if hasattr(record, "exc_info"):
-            new_record = copy(record)
-            new_record.exc_info = None
-            return super().format(new_record)
-        return super().format(record)
-
-
 class TTOptions(dict):
     """
     Use a dot "." to access members of dictionary
